requirements.txt
absl-py==2.3.1
altair==5.5.0
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
flatbuffers==25.9.23
gast==0.7.0
gitdb==4.0.12
gitpython==3.1.45
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.12.0
libclang==18.1.1
markdown==3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
ml-dtypes==0.5.4
namex==0.1.0
narwhals==2.13.0
numpy==2.2.6
opencv-python==4.12.0.88
opt-einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.1
pyarrow==21.0.0
pydeck==0.9.1
pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
streamlit==1.51.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
toml==0.10.2
tornado==6.5.2
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
werkzeug==3.1.4
wheel==0.45.1
wrapt==2.0.1
